+++ docs/source/exporters/onnx/package_reference/export.mdx
@@ -18,6 +18,8 @@ You can export models to ONNX from two frameworks in 🤗 Optimum: PyTorch and T
 
 [[autodoc]] exporters.onnx.main_export
 
+[[autodoc]] exporters.onnx.onnx_export_from_model
+
 [[autodoc]] exporters.onnx.convert.export
 
 [[autodoc]] exporters.onnx.convert.export_pytorch
